move user_scripts to tools and combine with math_models
