+++ gpkit/__init__.py
@@ -92,8 +92,8 @@ from .varkey import VarKey
 from .posyarray import PosyArray
 from .geometric_program import GeometricProgram
 from .signomial_program import SignomialProgram
-from .user_scripts import composite_objective, link
-from .user_scripts import make_feasibility_gp, closest_feasible_point
+from .tools import composite_objective, link
+from .tools import make_feasibility_gp, closest_feasible_point
 
 from .shortcuts import GP, SP
