--- gpkit/math_models.py
@@ -1,26 +1,0 @@
-"""Non-application-specific mathematical models"""
-
-
-def te_exp_minus1(posy, nterm):
-    """Taylor expansion of e^{posy} - 1
-
-    Arguments
-    ---------
-    posy : gpkit.Posynomial
-        Variable or expression to exponentiate
-    nterm : int
-        Number of terms in resulting Taylor expansion
-
-    Returns
-    -------
-    gpkit.Posynomial
-        Taylor expansion of e^{posy} - 1, carried to nterm terms
-    """
-    if nterm < 1:
-        raise ValueError("Unexpected number of terms, nterm=%s" % nterm)
-    res = 0
-    factorial_denom = 1
-    for i in xrange(1, nterm + 1):
-        factorial_denom *= i
-        res += posy**i / factorial_denom
-    return res
