+++ gpkit/tests/t_tools.py
