+++ gpkit/tools.py
@@ -1,11 +1,36 @@
+"""Non-application-specific convenience methods for GPkit"""
 import numpy as np
 
 from collections import Iterable
 
-from .nomials import Posynomial
 from .variables import Variable, VectorVariable
 from .posyarray import PosyArray
 from .geometric_program import GeometricProgram
+
+
+def te_exp_minus1(posy, nterm):
+    """Taylor expansion of e^{posy} - 1
+
+    Arguments
+    ---------
+    posy : gpkit.Posynomial
+        Variable or expression to exponentiate
+    nterm : int
+        Number of terms in resulting Taylor expansion
+
+    Returns
+    -------
+    gpkit.Posynomial
+        Taylor expansion of e^{posy} - 1, carried to nterm terms
+    """
+    if nterm < 1:
+        raise ValueError("Unexpected number of terms, nterm=%s" % nterm)
+    res = 0
+    factorial_denom = 1
+    for i in xrange(1, nterm + 1):
+        factorial_denom *= i
+        res += posy**i / factorial_denom
+    return res
 
 
 def make_feasibility_gp(gp, varname=None, flavour="max"):
